backend/requirements.txt
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
black==23.3.0
certifi==2024.12.14
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
Django==4.2.2
django-cors-headers==4.6.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.1.0
django-extensions==4.0
django-filter==25.1
django-rest-swagger==2.2.0
django-settings-export==1.2.1
django-webpack-loader==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
django-environ
Faker==18.10.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.4.3
MarkupSafe==3.0.2
msgpack==1.1.0
mypy-extensions==1.0.0
numpy
openapi-codec==1.3.2
packaging==24.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.2.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pydot==3.0.4
pyflowchart==0.3.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.5.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.24.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.13.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
